[CHORE] : FollowingAfterIdSearchActivity의 파일 정리
