+++ app/src/main/java/com/teambeme/beme/idsearchfollowing/adapter/IdSearchAdapter.kt
@@ -1,22 +1,22 @@
 package com.teambeme.beme.idsearchfollowing.adapter
 
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.teambeme.beme.R
 import com.teambeme.beme.databinding.ItemFollowingAfterIdsearchBinding
 import com.teambeme.beme.idsearchfollowing.model.ResponseIdSearchData
-import com.teambeme.beme.idsearchfollowing.viewmodel.IdSearchViewModel
 
-class IdSearchAdapter(private val idSearchViewModel: IdSearchViewModel) :
+class IdSearchAdapter :
     RecyclerView.Adapter<IdSearchAdapter.IdSearchViewHolder>() {
     private var idSearchDatas = mutableListOf<ResponseIdSearchData.Data>()
 
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int
-    ):IdSearchViewHolder {
+    ): IdSearchViewHolder {
         val inflater = LayoutInflater.from(parent.context)
         val binding: ItemFollowingAfterIdsearchBinding =
             DataBindingUtil.inflate(inflater, R.layout.item_following_after_idsearch, parent, false)
@@ -29,10 +29,10 @@ class IdSearchAdapter(private val idSearchViewModel: IdSearchViewModel) :
         holder.bind(idSearchDatas[position])
     }
 
-//    fun replaceIdSearchList(it: ResponseIdSearchData.Data) {
-//        idSearchDatas = it
-//        notifyDataSetChanged()
-//    }
+    fun replaceIdSearchList(list: MutableList<ResponseIdSearchData.Data>) {
+        idSearchDatas = list
+        notifyDataSetChanged()
+    }
 
     inner class IdSearchViewHolder(private val binding: ItemFollowingAfterIdsearchBinding) :
         RecyclerView.ViewHolder(binding.root) {
@@ -41,26 +41,26 @@ class IdSearchAdapter(private val idSearchViewModel: IdSearchViewModel) :
                 idSearch = idSearchData
                 executePendingBindings()
             }
-//            setBtnUnfollowClickListener(binding)
-//            setBtnFollowClickListener(binding)
+            setBtnUnfollowClickListener(binding)
+            setBtnFollowClickListener(binding)
         }
 
-//        private fun setBtnUnfollowClickListener(
-//            binding: ItemFollowingAfterIdsearchBinding
-//        ) {
-//            binding.btnFollowingUnfollow.setOnClickListener {
-//                binding.btnFollowingFollow.visibility = View.VISIBLE
-//                binding.btnFollowingUnfollow.visibility = View.INVISIBLE
-//            }
-//        }
-//
-//        private fun setBtnFollowClickListener(
-//            binding: ItemFollowingAfterIdsearchBinding
-//        ) {
-//            binding.btnFollowingFollow.setOnClickListener {
-//                binding.btnFollowingFollow.visibility = View.INVISIBLE
-//                binding.btnFollowingUnfollow.visibility = View.VISIBLE
-//            }
-//        }
+        private fun setBtnUnfollowClickListener(
+            binding: ItemFollowingAfterIdsearchBinding
+        ) {
+            binding.btnFollowingUnfollow.setOnClickListener {
+                binding.btnFollowingFollow.visibility = View.VISIBLE
+                binding.btnFollowingUnfollow.visibility = View.INVISIBLE
+            }
+        }
+
+        private fun setBtnFollowClickListener(
+            binding: ItemFollowingAfterIdsearchBinding
+        ) {
+            binding.btnFollowingFollow.setOnClickListener {
+                binding.btnFollowingFollow.visibility = View.INVISIBLE
+                binding.btnFollowingUnfollow.visibility = View.VISIBLE
+            }
+        }
     }
 }
