--- app/src/main/java/com/teambeme/beme/idsearchfollowing/model/IdSearchData.kt
@@ -1,6 +1,0 @@
-package com.teambeme.beme.idsearchfollowing.model
-
-data class IdSearchData(
-    val userID: String,
-    val userPicUrl: String
-)
