--- app/src/main/java/com/teambeme/beme/idsearchfollowing/model/RecentSearchData.kt
@@ -1,6 +1,0 @@
-package com.teambeme.beme.idsearchfollowing.model
-
-data class RecentSearchData(
-    val userID: String,
-    val userPicUrl: String
-)
