+++ app/src/main/res/layout/activity_following_after_id_search.xml
@@ -65,7 +65,7 @@
             android:layout_marginHorizontal="30dp"
             android:layout_marginTop="18dp"
             android:overScrollMode="never"
-            android:visibility="invisible"
+            android:visibility="visible"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
